fixes #241
This addresses @mattgarrish's issues regarding EPUB properties and separating out DC:Terms and improved organization.

If there is still an issue regarding "the information and resources section looks misplaced at the start of the document. Supplementary info like that is usually a trailing section or appendix. The crosswalk is the most important thing."
This can be copied into a new issue.